separated the classes Person and ClinicTime
