+++ .gitignore
@@ -1,0 +1,1 @@
+__pycache__/Person.cpython-39.pyc
