+++ ClinicTime.py
@@ -1,0 +1,6 @@
+class ClinicTime():
+    office_hours = list()
+    available_time = list()
+    taken_time = dict()
+    def __init__(self):
+        pass
